requirements_docker.txt
wsgiref==0.1.2
python==2.7.18
zipp==1.2.0
wheel==0.34.2
virtualenv==20.15.1
urllib3==1.21.1
typing==3.10.0.0
tornado==4.5.1
six==1.10.0
singledispatch==3.4.0.3
setuptools==44.1.0
scandir==1.10.0
requests==2.18.1
qrcode==5.3
pyyaml==3.12
pyutf8==0.1
pyusb==1.0.0
python-nmap==0.6.1
python-escpos==2.2.0
pyserial==3.3
platformdirs==2.0.2
pip==20.3.4
pillow==4.2.1
pathlib2==2.3.7.post1
olefile==0.44
importlib-resources==3.3.1
importlib-metadata==2.1.3
idna==2.5
filelock==3.2.1
distlib==0.3.8
contextlib2==0.6.0.post1
configparser==4.0.2
chardet==3.0.4
certifi==2017.4.17
backports-abc==0.5
argparse==1.4.0
argcomplete==1.8.2
appdirs==1.4.4
